core/generator: only upsert higher pending blocks
When upserting a pending block as the generator, only upsert if the
new block is at a higher height than the previous block. This wraps
up #519 for the generator. In addition, it should handle the situation
where two cored processes are both generating blocks as generator.
Whichever process first saves its pending block to the table wins. The
losing process will error and do nothing on that `makeBlock` call. If
the block hasn't been successfully committed yet, it'll pull it from the
pending block table and finish committing the block to the blockchain.